# api/swagger.yaml
swagger: "2.0"
info:
  title: RupeeStore API
  description: API for managing rupee store
  version: 1.0.0
basePath: /api/
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
    description: Bearer token for authentication
paths:
  /initiateSession:
    post:
      summary: user validation
      description: ruppee store authentication
      operationId: initiateSession
      x-swagger-router-controller: clientController
      consumes:
        - application/json
      produces:
        - application/json
      tags:
        - User
      parameters:
        - in: body
          name: Get initiateData
          description:  initiateData
          required: true
          schema:
            $ref: "#/definitions/initiateData"
      responses:
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /addCategory:
    post:
      summary: category adding
      description: ruppee store addcategory
      operationId: addCategory
      x-swagger-router-controller: adminController
      consumes:
        - application/json
      produces:
        - application/json
      tags:
        - RupeeAdmin
      parameters:
        - in: header
          name: Authorization
          type: string
          format: Bearer
          required: true
          description: Bearer token for authentication
        - in: body
          name: Get initiateData
          description:  initiateData
          required: true
          schema:
            $ref: "#/definitions/initiateData"
      responses:
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /getCategories:
    get:
      summary: Get all categories
      description: Get a list of all categories
      operationId: getCategories
      x-swagger-router-controller: clientController
      consumes:
        - application/json
      produces:
        - application/json
      tags:
        - RupeeClient
      parameters:
        - in: header
          name: Authorization
          type: string
          format: Bearer
          required: true
          description: Bearer token for authentication
      responses:
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /addOffer:
    post:
      summary: offer adding
      description: add rupee offer
      operationId: addOffer
      x-swagger-router-controller: adminController
      consumes:
        - application/json
      produces:
        - application/json
      tags:
        - RupeeAdmin
      parameters:
        - in: header
          name: Authorization
          type: string
          format: Bearer
          required: true
          description: Bearer token for authentication
        - in: body
          name: Get initiateData
          description:  initiateData
          required: true
          schema:
            $ref: "#/definitions/initiateData"
      responses:
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /getOffers:
    get:
      summary: Get offers
      description: Get a list of offers
      operationId: getOffers
      x-swagger-router-controller: clientController
      consumes:
        - application/json
      produces:
        - application/json
      tags:
        - RupeeClient
      parameters:
        - in: header
          name: Authorization
          type: string
          format: Bearer
          required: true
          description: Bearer token for authentication
      responses:
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /getOfferDetails:
    post:
      summary: Offer details
      description: details of particular offer
      operationId: getOfferDetails
      x-swagger-router-controller: clientController
      consumes:
        - application/json
      produces:
        - application/json
      tags:
        - RupeeClient
      parameters:
        - in: header
          name: Authorization
          type: string
          format: Bearer
          required: true
          description: Bearer token for authentication
        - in: body
          name: Get initiateData
          description:  initiateData
          required: true
          schema:
            $ref: "#/definitions/initiateData"
      responses:
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /getCouponcode:
    post:
      summary: coupon generate
      description: coupon for particular offer
      operationId: getCouponcode
      x-swagger-router-controller: clientController
      consumes:
        - application/json
      produces:
        - application/json
      tags:
        - RupeeClient
      parameters:
        - in: header
          name: Authorization
          type: string
          format: Bearer
          required: true
          description: Bearer token for authentication
        - in: body
          name: Get initiateData
          description:  initiateData
          required: true
          schema:
            $ref: "#/definitions/initiateData"
      responses:
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /GetPwaRewards:
    get:
      summary: Get Date
      description: Create a new trip with authentication
      operationId: GetPwaRewards
      x-swagger-router-controller: tripController
      consumes:
        - application/json
      produces:
        - application/json
      tags:
        - cheggout API
      parameters:
        - in: header
          name: Cookie
          type: string
          format: cookie
          required: false
          description: The cookie containing the ARRAffinity value.
        - in: header
          name: Authorization
          type: string
          format: Bearer
          required: true
          description: Bearer token for authentication

      responses:
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /GetPWAWalletPoints:
    get:
      summary: Get Date
      description: Create a new trip with authentication
      operationId: GetPWAWalletPoints
      x-swagger-router-controller: tripController
      consumes:
        - application/json
      produces:
        - application/json
      tags:
        - cheggout API
      parameters:
        - in: header
          name: Authorization
          type: string
          format: Bearer
          required: true
          description: Bearer token for authentication
        - name: Cookie
          in: header
          required: false
          type: string
          description: The cookie containing the ARRAffinity value.
      responses:
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /DeductWalletPoints:
    post:
      summary: Get Date
      description: Create a new trip with authentication
      operationId: DeductWalletPoints
      x-swagger-router-controller: tripController
      consumes:
        - application/json
      produces:
        - application/json
      tags:
        - cheggout API
      parameters:
        - in: header
          name: Authorization
          type: string
          format: Bearer
          required: true
          description: Bearer token for authentication
        - name: Cookie
          in: header
          required: false
          type: string
          description: The cookie containing the ARRAffinity value.
        - in: body
          name: DeductWalletPoints
          description: DeductWalletPoints
          required: true
          schema:
            $ref: "#/definitions/DeductWalletPointsData"
      responses:
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /RefundPoints:
    post:
      summary: Get Date
      description: Create a new trip with authentication
      operationId: RefundPoints
      x-swagger-router-controller: tripController
      consumes:
        - application/json
      produces:
        - application/json
      tags:
        - cheggout API
      parameters:
        - in: header
          name: Authorization
          type: string
          format: Bearer
          required: true
          description: Bearer token for authentication
        - name: Cookie
          in: header
          required: true
          type: string
          description: The cookie containing the ARRAffinity value.
        - in: body
          name: RefundPoints
          description: RefundPoints
          required: true
          schema:
            $ref: "#/definitions/DeductWalletPointsData"
      responses:
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
  /addTrip:
    post:
      summary: Create a new trip
      description: Create a new trip with authentication
      operationId: addTrip
      x-swagger-router-controller: tripController
      consumes:
        - application/json
      produces:
        - application/json
      tags:
        - Trip
      parameters:
        - in: header
          name: Authorization
          type: string
          format: Bearer
          required: true
          description: Bearer token for authentication
        - in: body
          name: trip
          description: Trip object
          required: true
          schema:
            $ref: "#/definitions/Trip"
      responses:
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /getTrips:
    get:
      summary: Get all trips
      description: Get a list of all trips with authentication
      operationId: getTrips
      x-swagger-router-controller: tripController
      consumes:
        - application/json
      produces:
        - application/json
      tags:
        - Trip
      parameters:
        - in: header
          name: Authorization
          type: string
          format: Bearer
          required: true
          description: Bearer token for authentication
      responses:
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /getTripById:
    post:
      summary: Get a trip by ID
      description: Get detailed information about a specific trip with authentication
      operationId: getTripById
      x-swagger-router-controller: tripController
      consumes:
        - application/json
      produces:
        - application/json
      tags:
        - Trip
      parameters:
        - in: header
          name: Authorization
          type: string
          format: Bearer
          required: true
          description: Bearer token for authentication
        - in: body
          name: tripId
          description: ID of the trip to retrieve
          required: true
          schema:
            $ref: "#/definitions/TripIdData"
      responses:
        200:
          description: Successful operation
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
...
  Trip:
    type: object
    properties:
      tripId:
        type: integer
        format: int64
      userId:
        type: integer
        format: int64
      tripType:
        type: string
      tripFrom:
        type: string
      tripTo:
        type: string
      tripDeparture:
        type: string
        format: date-time
      tripReturn:
        type: string
        format: date-time
      travellersNum:
        type: integer
        format: int32
      adults:
        type: integer
        format: int32
      childrens:
        type: integer
        format: int32
      class:
        type: string
      createdBy:
        type: integer
        format: int64
      updatedBy:
        type: integer
        format: int64
  TripIdData:
    type: object
    properties:
      tripId:
        type: integer
        format: int64
...
  initiateData:
    type: object
    properties:
      # uuid:
      #   type: string
      id_user:
        type: string
      bank_name:
        type: string
      bank_token:
        type: string
      # token:
      #   type: string
      ticket_module:
        type: string
      # last_login:
      #   type: string
      #   format: date-time
      # updatedAt:
      #   type: string
      #   format: date-time
  DeductWalletPointsData:
    type: object
    properties:
      rewardPoint:
        type: string
...
  ErrorResponse:
    type: object
    properties:
      code:
        type: integer
        format: int32
...